[ghc8] hsc: Remove alignment #let macro
The #let macro is invalid when cross-compiling a hsc file.
We already specify an alignment for the instances that we
define in this file, and we aren't cross-compiling across
architectures, so removing this line should be okay.

Future implentations of OpenXT that support alternative host
architectures might break without this, but we will likely
not be using Haskell for our toolstack by this time.

Signed-off-by: Nicholas Tsirakis <tsirakisn@ainfosec.com>